Refactor: Seperate Comment Item Widget
